Refactor prompt
