+++ src/cmd/cli/command/mcp.go
@@ -8,6 +8,7 @@ import (
 	cliClient "github.com/DefangLabs/defang/src/pkg/cli/client"
 	"github.com/DefangLabs/defang/src/pkg/login"
 	"github.com/DefangLabs/defang/src/pkg/mcp"
+	"github.com/DefangLabs/defang/src/pkg/mcp/prompts"
 	"github.com/DefangLabs/defang/src/pkg/mcp/resources"
 	"github.com/DefangLabs/defang/src/pkg/mcp/tools"
 	"github.com/DefangLabs/defang/src/pkg/term"
@@ -73,6 +74,10 @@ set_config - This tool sets or updates configuration variables for a deployed ap
 		// Setup resources
 		term.Debug("Setting up resources")
 		resources.SetupResources(s, &providerID)
+
+		//setup prompts
+		term.Debug("Setting up prompts")
+		prompts.SetupPrompts(s, &providerID)
 
 		// Setup tools
 		term.Debug("Setting up tools")
