+++ src/pkg/mcp/prompts/common.go
@@ -1,0 +1,8 @@
+package prompts
+
+func getStringArg(args map[string]string, key, defaultValue string) string {
+	if val, exists := args[key]; exists {
+		return val
+	}
+	return defaultValue
+}
