+++ src/pkg/mcp/prompts/prompts.go
@@ -1,0 +1,12 @@
+package prompts
+
+import (
+	"github.com/DefangLabs/defang/src/pkg/cli/client"
+	"github.com/mark3labs/mcp-go/server"
+)
+
+// SetupPrompts configures and adds all prompts to the MCP server
+func SetupPrompts(s *server.MCPServer, providerId *client.ProviderID) {
+	//AWS BYOC
+	setupAWSBYOPrompt(s, providerId)
+}
